Add files via upload
This is the final project website for a base level fundamentals of Web Dev class. it allows for users to find the brand of their machine and then see the troubleshooting steps to fix their machine. For class purposes, i did not include every single model for each brand, and I did the very basic trouble shooting guides to show off my design creativity.